Split cloud backup manager logic
Moves restore-all and other-backups supervisor flows into dedicated modules, and splits enable/passkey handling into smaller focused files. This is a structural refactor only; behavior stays the same.
